# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Passagens",
...
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8044",
            "description": "Servidor API de Produção"
        }
    ],
    "paths": {
...
        "/api/flights": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Listar todos os voos",
                "operationId": "4baffc280d4c3e239575b8c57e979397",
                "responses": {
                    "200": {
                        "description": "Lista de voos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Flights"
                ],
                "summary": "Criar um novo voo",
                "operationId": "429a30d2ba9b4c386fd85eefd684225e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin_airport_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "destination_airport_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "flight_number": {
                                        "type": "string",
                                        "example": "FL123"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voo criado com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flights/{id}": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Mostrar um voo específico",
                "operationId": "6ab7311543197daa8b1ff4be0f88c90e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do voo"
                    },
                    "404": {
                        "description": "Voo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Flights"
                ],
                "summary": "Atualizar um voo",
                "operationId": "29d9b985eff7a7d40f56bd4a377854c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "origin_airport_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "destination_airport_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "flight_number": {
                                        "type": "string",
                                        "example": "FL123"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voo atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro na validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Flights"
                ],
                "summary": "Deletar um voo",
                "operationId": "fe823c1af305d4b2566476b298c44e5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voo deletado com sucesso"
                    },
                    "404": {
                        "description": "Voo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flights/{id}/passengers": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Listar passageiros de um voo",
                "operationId": "1f5b32a39d536208f122088133c9e559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de passageiros"
                    },
                    "404": {
                        "description": "Voo não encontrado"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flights/search": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Search for flights",
                "description": "Search for flights based on criteria",
                "operationId": "searchFlights",
                "parameters": [
                    {
                        "name": "origin_airport_id",
                        "in": "query",
                        "description": "ID of the origin airport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "destination_airport_id",
                        "in": "query",
                        "description": "ID of the destination airport",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date of the flight",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price of the flight",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token para autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}